requirements.txt
alembic==1.7.3
asgiref==3.4.1
bson==0.5.10
click==8.0.1
fastapi==0.68.1
Faker==8.14.0
greenlet==1.1.1
h11==0.12.0
httptools==0.2.0
jsonify==0.5
Mako==1.1.5
MarkupSafe==2.0.1
pydantic==1.8.2
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==0.19.0
PyYAML==5.4.1
pytest==6.2.5
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.25
starlette==0.14.2
typing-extensions==3.10.0.2
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
websockets==10.0
python-jose==3.3.0
passlib==1.7.4
PyJWT==1.7.1
bcrypt==3.2.0
fastapi-jwt-auth==0.5.0
colorama==0.4.4







